[skottie] Visible track matte adjustments
AE allows controlling layer visibility independently of whether it is
being used as a track matte.

The Lottie 1.0 spec is aligned with AE, and defers layer visibility to
the 'hd' property.

Legacy Bodymovin uses a separate property to flag track mattes, 'td',
and assumes it also implies the layer is hidden.  It does not export
an 'hd' property for track mattes at all.

Some assets in the wild appear to be mixing these semantics as they
both flag mattes with 'td' and also mark them as visible with
'hd': false.  The expectation is that these mattes are not hidden.

This CL updates Skottie to always observe an explicit 'hd' property,
and only consider legacy 'td' when visibility is not otherwise
specified.

Change-Id: Id2bfdcdb7182518dc062b2748127344e90cdc31c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/945156
Reviewed-by: Brandon Wilson <bwils@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>