[graphite] Add API to return Android-specific keys back to the user
In order for clients to be able use the Android-specific API they, at some point, need to be notified about new keys. This new API allows clients to supply a callback function that will be invoked whenever a new key is added to the GlobalCache.

This CL just adds the API and updates DM a bit. The callback will be hooked up to Graphite's internals in a follow up CL.

Bug: b/390186667
Change-Id: I6ce6603af56f6dbd1f0ad3d15aaeef26f76d0276
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/944076
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>