[Graphite] Update Vulkan buffer barriers.
The biggest issue is that we didn't put in any sort of barrier after
doing a buffer to buffer copy. In the Vulkan backend the only time we
really would have hit this today is when initializing the static draw
buffers.

Also did just a little cleaning up while I was in this code.

Change-Id: Ia26f3d9dc7d9e11e9554cc21a7eb435820f2dfe6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/944977
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>