[pdf] Do not scale miterLimit
In "[pdf] Reduce drawing glyphs as paths" [0] a new function
`scale_paint` was introduced to reduce redundancy in the fonts that need
to be embedded in the PDF. This function properly scaled the
`SkPaint::fWidth` but it also incorrectly scaled `SkPaint::fMiterLimit`.
The miter limit is already a ratio and so should not be scaled. Remove
this incorrect scaling.

This issue was introduced partially due to a lack of understanding of
the units of the miter limit and that it is a ratio and not the max
miter length. Update the documentation on `SkPaint::setStrokeMiter` to
include some actually useful information.

[0] https://skia.googlesource.com/skia/+/8884cb1822618b0ce95a25e2ef1560c98fa8b081

Bug: 379510433
Change-Id: Idab1074a4a41938f32c38b148e48b28981dd7c95
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/946917
Reviewed-by: Brett Somocurcio <brettos@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>