[pdf] Developer build dump of input structure tree
When debugging PDF AX issues it is sometimes difficult to tell if Skia
might be leaving information out of the PDF's structure element tree
(usually because nothing was drawn for that element) or if the
information was never provided to begin with. This change adds some
normally compiled out code to allow a developer to easily create a build
which dumps the structure as it was provided to Skia. This has been
quite helpful in several recent issue investigations.

Change-Id: I8da08cf37457d20f3d8ebd42ccc446e55ec0db06
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/946918
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>