fix image serialization for MSKPs on Android using derived generators
Captured MSKPs fail to serialize images because our Android factories create SkImages with DeferredFromGenerator. This returns a SkImage_Lazy that doesn't have a way to getPixels.

This CL 1) Changes the android factories to make SkImage_LazyTextures and 2) allows our sharing proc to accept a GrDirectContext which is needed to rasterize LazyTextures for MSKPs

Change-Id: Id12c71c9e2ef8e7010a9fec2c74be8be1a2f89d0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/946217
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Jorge Betancourt <jmbetancourt@google.com>