Correct "proxy" terminology, remove setProxy
The commonly accepted terms of the "proxy pattern" are that there is a
Proxy (SkTypeface_proxy) which implements the Subject (SkTypeface)
interface. The Proxy maintains a reference (realTypeface) to the
substituted object which is of a concrete type RealSubject which also
implements the Subject interface. The current naming convention in the
code is the reverse of this convention, so correct the terms and clarify
where needed.

This also removes SkTypeface_proxy::setProxy and rolls it into the
constructor where it belongs.

Change-Id: I93ecaa70b605c9c100fdb490bf9087780192acbb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/947658
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>