SkTypeface_fontconfig::onMakeClone to keep pattern
Fix the long standing TODO that FC_MATRIX and FC_EMOBOLDEN need to be
passed on through clones. Do so by having onMakeClone return a
SkTypeface_fontconfig with a reference to fPattern. Correct the style by
keeping any weight or width offset and slant override.

Note that this fixes onMakeClone, but does not fix the similar issue
(and TODO in onGetFontDescriptor) with serialization and
deserialization.

Change-Id: I4869d886b2c9bb2ae4df519488d92d4e5adb147b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/948176
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>