Small MaskFilter cleanups
hasNoBlur()
This started out as looking odd to do a multiply and then
a comparison instead of doing a comparison to a smaller number.
It wasn't clear to me why 1/3 was special, so I did some research
through git history and some math to figure it out. That threshold
was a bit pessimistic, AFAICT, so I used the "correct" value and
have it hard-coded which saves us 2 multiplies.

draw_rrect_into_mask()
I noticed some extra move calls as I was stepping into this function
and realized we weren't passing by reference.

SkIPoint->SkIVector
I was confused why we were looking for a point and then I realized
we were looking for a dx, dy which feels better (in my brain) inside
a vector-named type.

SkScalar->float
We've been occasionally doing this throughout the codebase because
SkScalar can't be anything else anymore.

SkBlurMask::BoxBlur()
The comment here really threw me off. I figured out what it meant
and rewrote it to hopefully be less confusing.

Change-Id: I60c76fbabe7aa1f3117da82ade71979b5a8e5b6a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/947636
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>