[Graphite] Fix ResourceCache Token if size changes.
For resources with zero size, we always set the UseToken to be the max
value as a special signal. However, we now allow us to change the
resources size when we return it to the cache. This can cause a
resource to go from size 0 to greater zero which invalidates the
UseToken value on it. Thus it then needs to call
setResourceUseToken.

Previously we only updated the Resource UseToken if we were purgeable.
That meant if we were non-purgeable we would exit returning the
resource which would then try to validate the cache which the now
non-zero sized resource had an invalidate token.

This change now has us always set the UseToken when returning a
resource. I don't believe there is any reason why we can't set it
before a Resource is purgeable. It just won't have much effect besides
keeping our assumptions in a valid state.

Change-Id: I35e4bbc6f80cf4d1f5268b3c9a62ef3ab1cf6476
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/948916
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>