Serialize stableKeys, rather than SkSL code, for runtimeEffects in skp files
When I added StableKeys I intentionally didn't serialize them in skp files, hoping that it wouldn't matter (since the SkSL would be written anyways - so the generated images would look the same).

With the addition of the Android-style Pipeline keys this does become a testing issue. When rendering the skps with one of the gr*testprecompile configs the appearance of SkSL that should be stably keys is causing some of the saved skp files to fail. In all cases this is due to the usage of the SkLumaColorFilter.

At a meta-level, serializing the stable keys better preserves the information and intent that Skia possesses. It does introduce some additional fragility to SKP files but that should be mitigated, somewhat, by the infrastructure around Android-style Pipeline key versioning (i.e., if the Pipeline key version needs to be bumped bc of stable keying changes we can also bump the SKP version).

This change is tested via the serialize-8888 jobs, specifically:

 --config serialize-8888 --match composeCF async_rescale_and_read_alpha_type

Bug: b/394827799
Change-Id: Id6ce6cf78d989aa4252036918a9010971e216023
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/947956
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>