Update rewrite_includes.py to support duplicate header names
Collects short names to a list of include paths, allowing duplicates
to be represented. Updates the rewrite logic to be okay with duplicate
headers as long as the actual include directive has already been made
explicit. Still supports automatically rewriting paths that have a
single path match, as well as sorting.

If a path cannot be rewritten, it prints the ambiguous cases that
should be manually chosen but otherwise does not modify the problematic
line of code. This will have it exit with an error status so should
work fine with the presubmit hooks as well.

I tested manually by modifying src/gpu/graphite/Device.cpp to just
include Device.h w/o a base path (becoming ambiguous with Device.h
in src/gpu/ganesh). With the duplicate restriction removed, I also
simplified the ignore list (leading to some rewrites) and removed
the _graphite suffix on some types.

I opted to keep headers with the _Graphite suffix for types that are
participating in a class hierarchy defined in core/.

Change-Id: I7bd84e996087dc701f48afd9fca8c7c19d54d28d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/949536
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>