Support ISO gainmap colorspace in PNG
The gainmap colorspace math is not encoded in the ISO payload, it's expected to
be on the ICC profile of the gainmap. So, encode an ICC profile on the gainmap
image for 3D gainmaps. For 1D gainmaps, don't encode a profile and just drop
the gainmap colorspace on the floor.

Also, add tests, for:
* Decoding a known-good image
* Decoding malformed pngs with a gainmap
* Round trip encoding ane decoding an image

Bug: b/395892646
Change-Id: I090894ac87f43d14c0dfea2e7dd6affa8d46f3b4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/950557
Commit-Queue: Alec Mouri <alecmouri@google.com>
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>