Remove unnecessary parameter in blur mask code
Found this was investigating another change. margin is an output
parameter, but since the result is not used after, we can save
a copy operation by passing in nullptr.

Note we *do* need margin in the earlier calls in these functions
to establish how much to expand the mask when blurring.

Change-Id: I4290fe14ee8e5afae1f23a77a67a4951dbb4b13e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/952217
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Daniel Dilan <danieldilan@google.com>