[graphite] Remove TextureInfoData and TextureSpec in TextureInfo
The intermediates, TextureInfoData and its subclasses, and the backend
structs [Foo]TextureSpec are removed.

A private TextureInfo::Data virtual class is introduced that holds the
sample count and mipmapped state, with the backend [Foo]TextureInfo
classes extending. This exposes those fields and preserves their field
layout before this CL, and also makes them equivalent to the removed
TextureSpec structs.

Since TextureInfo::Data is private to TextureInfo and can be friended,
the virtual functionality that had been on TextureInfoData is declared
up front to simplify the number of types that have to be implemented.

In situations where there is a backend context (e.g. a Caps object)
or calling a function from a backend-specific compilation unit,
template traits are used instead of increasing the number of virtual
functions. With a few more follow up CLs, the only virtual
functionality on TextureInfo::Data will be related to equality and
initialization/assignment operators.

I opted to have the template functionality be part of the
FooTextureInfo subclasses directly, instead of having a specialization
for such as `TextureInfoTraits<FooTextureInfo>`. If it was that way,
it would require all the callers/users of the templated functionality
to include multiple headers and if the traits specialization wasn't
included, the compiler error messages were pretty opaque.

Lastly, this goes through all the backends and updates them to
access their backend data directly using TextureInfoPriv::Get<T>,
which just casts the underlying SkAnySubclass.


Change-Id: I5f58146305175ca4da4d0feca5d8499cc850da35
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/952676
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>