Remove transform_scanline functions from SkJpegEncoderImpl
These can be done generically using the SkConvertPixels function and SkRasterPipeline. This makes it simpler for adding color types and is a step towards completely removing transform_scanline functions.

This CL replaces the fProc field in SkJpegEncoderMgr, and keeps std::optional values of the src and dst SkImageInfo with height of 1, if a color type conversion is necessary. Then calls SkConvertPixels for every line.

Bug: 40041526
Change-Id: I1b71c0fa5cfd1695a30ec29e2c94b0d659d791d6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/953517
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>