[graphite] Add user-defined known runtime effects to ShaderCodeDictionary
We need some object to hold both the ShaderSnippets and SkRuntimeEffects for the user-defined known runtime effects that is also universally accessible - enter the ShaderCodeDictionary.

I've structured this so the ShaderCodeDictionary gets the SkRuntimeEffects in its ctor so we don't have to guard threaded access.

This is the meat of the new feature. The API is just a thin shim on top of this while the unit test is a monster.

Bug: b/390186667
Change-Id: I46b556075b493a036724d09bb6f67aa693d75c89
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/953358
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>