Fix ninepatch size calculations when dealing with fractional radii
This adds documentation and also produces a ninepatch that is
the optimal size (had previously been a bit bigger than necessary
and the center was not properly calculated).

In the fractional radii case, this is much better because the
blurring is now symmetrical and the right size.

In the rrect_blurs GM this improves the second case but makes the
other 3 even more different in terms of raw pixels diff.

Control:
70 pixels diff
362 pixels diff
222 pixels diff
116 pixels diff

With this cl:
112 pixels diff
265 pixels diff
347 pixels diff
148 pixels diff

I think that's because of b/397749778. With the mask being
smaller, there's more for the blitAntiH to draw on top and
bottom, causing there to be more pixels different.

This causes a diff in simpleblurroundrect (which I cleaned up a bit
while poking around) because row 1, pair 0 now can be drawn with the
fast path. The (adjusted for scale) rrect is 75 x 75 pixels wide,
with a corner radius of 3. The blur margin is calculated to be 27.
Under the old math, the minimum nine-patch size was 117
(27 * 4 + 3*2 + 3) which was bigger than 75, so it fell back to the
slow path. Now, the nine-patch minimum size is 61 (27*2 + 3*2 + 1)
so the fast path can be used.

What's interesting about the simpleblurroundrect diff is that it's
only on RGB_565 and BGRA_8888 (as of this tryjob), so I wonder
if there's something special with those blitters that's causing
the inconsistency? SkARGB32_Black_Blitter is used on RGB_8888
which does not appear to have the problem seen in b/397749778

For reference: original CL to add this impl
https://codereview.chromium.org/48623006/

Change-Id: I717b7f303b804372f26d48e04b4caa6b4b24c9c6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/953496
Commit-Queue: Florin Malita <fmalita@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>