[graphite] Append new clip cache entries to end of list.
Before when we had multiple clip masks with the same shapes but
different device bounds we'd add the new ones to the second position
(the first is the hash table entry itself). Now we add them at the end,
which means that they will end up being sorted from smaller device
bounds to larger, and when searching we'll end up using the clip mask
with the smallest valid bounds.

Bug: b/379148010
Change-Id: Iaafdef9f3351103c5d429b541adc9e2cd0e0270b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/955036
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>