Make ARGB32 blitting more consistent on x64 and NEON (arm) codepaths
Skia has a few ways to approximate dividing by 255 (e.g.
alpha transparency) by dividing by 256 and doing some corrective
factors. We should strive to make our use of them consistent, at
least within a single blit or with blitters that are used together.

Both blitAntiH and blitMask mix a src color onto a dst color using
an alpha value to control the src color (it should combine with whatever
alpha src already has). These should be consistent with the color
they produce, even if our different blitters differ.

This CL fixes several inconsistencies:
 - The premultiplied color in SkARGB32_Blitter was calculated in
   a non-standard way, which meant blitMask was using a different
   color sometimes than blitAntiH.
 - The neon accelerated path for blitting masks used different
   math for the 8-wide code vs the leftover at the end. This,
   in part, seems to be a result of [1] making the SkBlendARGB32
   function work better when blending a color and a translucent
   version of itself. Now masks on ARM devices should be consistent
   across the entire mask. I also changed some of the variable names
   to make it easier to compare the implementation (it took me a
   while to spot the key differences).
 - The SIMD (skvx) implementation of blit_row_color32 was using
   a more precise way to blend colors. This was another discrepancy
   between the two methods on x64 devices.

Maybe in a follow-up CL, I can try making the neon and/or skvx
blit mask code more accurate (especially for the same color) but
I'm a bit concerned about performance implications to do it in this CL.
Note to future me: [2] is a historical CL that was doing the same thing.

While testing this out, I found a broken gm (blend.cpp). It was
supposed to produce 3 squares of the same color, but it did not
at tip of tree, nor after my change. After my change, however,
the gm started failing on serialize-8888 because calls to
saveLayerAlpha followed by drawRect get merged together, but in
a normal 8888 backend, the opaque pixels are drawn into a pixmap,
and then blitted to the screen using Sprite_D32_S32 (an
SkSpriteBlitter) which uses blit_row_s32a_blend which uses
SkBlendARGB32. SkARGB32_Blitter::blitRect, in contrast, uses
SkBlitRow::Color32 on the premultiplied color and this yields
slightly different results. I'm not sure if that's worth addressing.
In the short term, I deleted the gm. If it's important, we should
write it as a unit test so it doesn't fail silently for years (?).

In any case, the main goal of the linked bug is achieved.
The rrect_blurs gm has many fewer diffs than before on the 4
test cases. To confirm, I patched this into Chromium and removed
SK_IGNORE_BLURRED_RRECT_OPT. The major pixel diffs have gone away.

[1] https://codereview.chromium.org/2097883002
[2] https://codereview.chromium.org/1092433002

Change-Id: I7ff847ecebc1be5626d0bced43b3c8906821d67f
Bug: b/397749778
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/953516
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>