[infra] Add iPhone15Pro jobs
These are still failing with: xcodebuild: error: Unable to find a
destination matching the provided destination specifier

As I understand it, this is out of our hands and we're just waiting for
it to be fixed for us. I'd like to get this landed and just let these be
red to prevent merge conflicts when the fix actually rolls out.

Bug: b/382573337
Change-Id: I06dbdcf1e76cdc835b7c0f659294fe205ddc7920
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/928741
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>