[graphite] Add universal TextureFormat enum
TextureFormat represents the union of backend formats that we care to
support. It adds utility functions for blocks-per-byte, name,
compression type, etc. that currently exist in near duplicate for
Metal, Vulkan, and Dawn.  As TextureFormat becomes the primary way for
core code to represent texture data, these functions will be cleaned
up and/or moved to be Ganesh-only.

The documentation in TextureFormat outlines the plan for mapping to
and from an SkColorType, although the actual implementation of that
will come in a later CL. The other major architectural change that
this CL enables is reducing AttachmentDesc to a format and sample
count instead of requiring an entire TextureInfo. Besides being
lighter weight, this makes the definition and serialization of
RenderPassDescs more backend agnostic.

The bulk of this CL is defining and documenting TextureFormat, and
adding it to TextureInfo. Its used in a few minor places for now:
 - attachment label generation
 - getting channel masks for yuv texture info w/o requiring Caps
 - removing some low-hanging backend calls in Vulkan

Bug: b/397666606
Change-Id: Icd8ae9ed7dd1cefe238c7667996a52145d19bd13
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/952736
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>