Remove and de-duplicate Android FontMgr streams
The Android FontMgrs have SkTypefaces which hold onto resource streams.
However, these are not needed as the proxy already is holding onto these
resource streams. Also, these resource streams are not currently
de-duplicated. De-duplicate them.

Currently the user preference for "isolated" resources is ignored and is
treated as if always true. This commit does not fix that.

Bug: 401319002
Change-Id: I7d6fa242f9b0fda15c8b8fbe7c4e7e1135671293
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/960216
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/969097