[skif] Fix device-to-layer bounds mapping
If there was perspective in the canvas matrix that mapped
device-space Z values very far from 0, the inverse mapping
from device to layer space was very inaccurate when using
the 4x4 inverse and an assumed z value of 0.

This switches to using the inverse of the flattened 3x3
for bounds mapping, which avoids this issue. The 4x4 has
to be kept in order to reconstruct relative transforms
between layers.

Bug: b/401088844
Bug: b/401085912
Change-Id: I1dd4885bc691d0a730e555bd56422af62d2fcf34
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/960356
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>