Fix need query for copyOnWrite for dual-proxies Ganesh images.
The ProxyChooser class picks the volatile vs stable copy proxy based on
the number of render tasks added to the volatile proxy since creation.
However, if we are drawing the snapped image into the same surface,
in device_DrawTexture we get the src view before we actually create the
new task for the draw image. Thus the Chooser thinks it is still safe
to return the volatile proxy for the src.

This change updates the ProxyChooser surfaceMustCopyOnWrite call,
which is made earlier, to also check if the volatile proxy matches
the surface.

Bug: b/387306744
Change-Id: Ib8ab2692ef303252ddd4b0798b1b6cb3caf52862
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/960616
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>