Move definition of SkSVGDOM::MakeFromStream to cpp file
This change moves the definition of SkSVGDOM::MakeFromStream from
SkSVGDOM.h to SkSVGDOM.cpp.

SkSVGDOM::MakeFromStream is exported but its boty is inline defined
in SkSVGDOM.h which calls SkSVGDOM::Builder::make which is not
exported. This causes link error of undefined symbol of
SkSVGDOM::Builder::make if Skia is built as a dynamic library.

Bug: 401979813
Change-Id: I8c5785fd61df4b4eb7aff08403181f6829d28447
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/960897
Commit-Queue: Florin Malita <fmalita@google.com>
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>