[graphite] Possible fixes for ClipAtlasManager PGO crash.
In ClipAtlasManager::evict() we can access the member of a deleted
object at the bottom of a loop, so added a break; to avoid that.

It's possible that PGO might be optimizing out the second
"entryList = fMaskCache.find(maskKey);" call in findOrCreateEntry(),
so using a different variable name to make it clear.

Bug: chromium:399071973
Change-Id: I604e77c13ed53fe918bef5355dcc3c2df8977f41
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/961256
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>