[graphite] Do not unpremul color data in gradient storage buffer
This was likely a copy-pasta error from the color processing in the
texture proxy gradient case. That used unpremul colors because it was
setting data per pixel using the SkBitmap data that would then
re-premul it.

The gradient color stop transformer has already prepared everything
for direct use, so there's no need to change it further when writing
to the gradient storage buffer.

Bug: b/401546700
Change-Id: I9687de7959a964fbf255d89a7dd5b47f3b36af7c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/962280
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>