[pdf] Correct /Parent for outline entries
Before this change the /Parent of an outline entry incorrectly used an
indirect reference to itself instead of its parent. Fix this by properly
assigning the indirect reference to the top level outline dictionary and
emit an indirect reference to the parent for /Parent instead of an
indirect reference to the child.

Bug: 383706655
Change-Id: Ibbf07098b889795c0970fef00a83e75445e9cb05
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/962778
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>