SkJpegMetadataDecoder: Fall back to base color space, not sRGB
The ISO 21496-1 specification has changed the default behavior for when
it is indicated to use the alternate image color space for the gain map
application color space, but no alternate image color space is
specified.

Prior to this, one was to use sRGB. Now, one is to use the base image
color space.

Bug: chromium:402033761
Change-Id: Ie9eb9c88de0f38bc345666e2bf79f26a9df13cd4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/962756
Reviewed-by: Christopher Cameron <ccameron@google.com>
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Commit-Queue: ccameron chromium <ccameron@chromium.org>