Fix CoreText data font weight calculation
In macOS 15 a change was made so that a font weight of 0 is treated like
a font weight of 1. This breaks automatic discovery of the conversion
from CoreText [-1, 1] weights to OpenType [1-1000] weights. In
particular, when converting the CoreText values to OpenType values, the
OpenType values 0 through 50 are all off by one. To fix this, get the
conversion associated with weight 11 and then back project that from the
conversion associated with weight 100. The number 11 was chosen because
CoreText treats font weights 1-10 as mapping to 100-1000 as a special
backwards compatibility for OS/2 version A.

Change-Id: I049bdc84d4c3f79a8659549fa836d1608aa5c65c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/964916
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Eric Boren <borenet@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Reviewed-by: Eric Boren <borenet@google.com>