[graphite] Intersect scissor and not viewport with attachment
https://github.com/gpuweb/gpuweb/issues/373 is resolved, so intersecting
the viewport with the attachment bounds in Dawn is no longer necessary.

On the other hand, SetScissorRect takes in unsigned integers [0], as
does the equivalent API in Metal [1]; intersecting the Graphite scissor
(which can have negative coordinates) with the attachment bounds ensures
we don't overflow.

[0] https://www.w3.org/TR/webgpu/#dom-gpurenderpassencoder-setscissorrect
[1] https://developer.apple.com/documentation/metal/mtlscissorrect?language=objc

Bug: b/378729995
Change-Id: I7d3c5edc1b61f5a2ec20628e037ff88c59275d61
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/961617
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>