Move Mac11 jobs onto Mac14 machines
The skolo's 2 MacMini9.1 machines (running Mac11) will be decomissioned
soon. We've been allocated 8 similar machines (running Mac14)
from another pool [1] so this transfers the Mac11 jobs
to them.

I made a few of the jobs that had been conditionally on the CQ
always on the CQ (since we have the capacity for it)
 - Test-Mac14-Clang-MacMini9.1-GPU-AppleM1-arm64-Debug-All-ASAN_Graphite_Dawn_Metal
 - Perf-Mac14-Clang-MacMini9.1-GPU-AppleM1-arm64-Release-All-Graphite_Dawn_Metal
 - Test-Mac14-Clang-MacMini9.1-GPU-AppleM1-arm64-Debug-All-Graphite_Native_Metal

and added the following job to always be on the CQ
 - Test-Mac14-Clang-MacMini9.1-GPU-AppleM1-arm64-Debug-All

Test-Mac11-Clang-MacMini9.1-CPU-AppleM1-arm64-Debug-All-NativeFonts
was not able to be fully replaced because the OS is different,
but there was already a Mac14 version of that job.

[1] https://chromium-swarm.appspot.com/botlist?c=id&c=task&c=device_type&c=gpu&c=lastSeen&c=os&c=status&d=desc&f=pool%3ASkia&f=mac_model%3AMacmini9%2C1&k=mac_model&s=status

Change-Id: I861f5eaaaee8b02bc210a015b1e332a1cac8ead8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/966925
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>