[graphite] Fix swapchain creation for BGRA_SRGB
VulkanCaps is updated to include a rendering-only definition for
BGRA_sRGB so that wrapping the backend texture as an SkSurface succeeds.
I chose to explicitly disable picking an SRGB format in the
GraphiteNativeVulkanWindowContext because its use requires setting the
dst colorspace to be linear sRGB. This doesn't interact well with
Viewer's UI controls to switch between legacy and color managed modes.

As skgpu::TextureFormat gets used more for the format configuration, I
think this will be cleaned up and readback can be supported with CPU
swizzling.

On my linux workstation, the first format returned by Vulkan is
B8G8R8A8_SRGB. Before
https://skia-review.googlesource.com/c/skia/+/952736 this would return
false from vkFormatIsSupported(), but afterwards it was being mapped to
TextureFormat::kBGRA8_sRGB so the GraphiteNativeVulkanWindowContext was
assuming it was renderable. Eventually, the backend texture wrapping
would fail for the swapchain and initialization would fail. During the
cleanup post-fail, some other small issues caused it to crash hard:

1. At the time of cleanup, the present queue hadn't been created yet so
it would segfault while waiting on it in destroyContext().

2. If the buffers failed, we wouldn't always propagate the failure out
of createBuffers() and then eventually we'd hit a later problem post
initialization.

3. There is still an unfixed issue where deleting the VMA object is
segfaulting inside the linux radeon drivers, but this only happens if
init was aborted. It works fine quitting viewer after a successful init.

Change-Id: I53eca3e2c5734605adf9af5b7e7d8f3bcfb1f086
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/967556
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>