Implement asImageFilter for SkTableMaskFilter
For this GM, the table mask filter affects coverage==0. So for the Graphite backend, all pixels in the clip will be affected which we determined more correct behavior than just pixels in the bounds box (which is the result for the other backends). Reasoning being that pixels inside the bounding with zero coverage don't belong to that contour more than any other pixel.

Bug: 304823483
Change-Id: Ibbfe7581cb717815b485ac471b8ed4e9922430e0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/963836
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>