[graphite] Add ProxyCache support to RasterPathAtlas.
This involves some changes to how we generate UniqueKeys for path
Shapes. Before we wouldn't allow a key for empty or volatile paths.
However, an empty path is semantically the same as an empty Shape.
And the only reason we don't allow volatile paths is because we
don't want them in long-term caches. But we can just check for
isVolatile() in that case.

Bug: b/404880920
Change-Id: If884ec5387fd13553a50d073be90e9b5c74b7c61
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/969101
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>