Decouple SkEdge and SkAnalyticEdge
This will allow some things in SkEdge (the non-AA version) to be a
bit simpler and allow any refactorings or changes in behavior in
that area to *not* impact our anti-alias code, which can have a rather
wide blast radius in terms of updating other peoples' tests.

Note that this also removes the sortY param in several functions
as it was always true. It's a leftover from the DAA code [1] which
has long since been deleted.

[1] https://skia-review.googlesource.com/c/skia/+/31940/4/src/core/SkScan_DAAPath.cpp

Change-Id: Ibb77861c08f6129a26fcaaa660f5b7d9c955fa86
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/972729
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>