[graphite] Expand PrecompileContext::reportPipelineStats
We want to re-use this entry point to determine a reasonable Pipeline cache size. The new kPipelineCache statistics should allow us to gauge how large the Pipeline cache would need to be to accommodate X minutes of Pipelines (where X is the interval between calls to reportPipelineStats).

This does introduce a bit of overhead even when not used (i.e., 16b/Pipeline, an assignment/new Pipeline and a comparison on Pipeline lookup). We could only compile it when SK_HISTOGRAMS_ENABLED is defined but that would make testing difficult.

Bug: b/358074434
Change-Id: I73cb0fe981eed9f275ccc459879c34f84bff2921
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/971176
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>