Implement asImageFilter for SkShaderMaskFilter
The former implementation of SkCanvas::onDrawEdgeAAImageSet2 did not account for mask filters being applied in the restore layer as an image filter. This CL changes that function so that we have a new layer for each image in the imageSet, so that we can apply the image in the paint as a shader and correctly use our mask filter in Graphite.

Bug: 304823483
Change-Id: I2159913d6f799e8b80b763fbebc2f6e1d39ff210
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/970476
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>