[Fontations] Path not modified for advanced metrics, PostScript name
Don't set path as modified. Path being modified is meant for emboldened
or otherwise filtered or geometrically altered path contours. That's not
the case when the font size is different from 1. SkScalerContext
parent implemention still takes care of setting the path modified flag
to true if the Fontations-returned path is to be emboldened.

Add storing the PostScript name to the advanced metrics struct while
we're at it.

Locally tested with steps from issue chromium:397880053.

Bug: skia:407716852, chromium:397880053
Change-Id: Ia02c2288490e5ff89e31a709cb1a06c7a9cd5492
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/973517
Commit-Queue: Dominik Röttsches <drott@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Dominik Röttsches <drott@google.com>