Simplify "font with random glyph mask formats" gm.
Remove the use of forceGenerateImageFromPath. This was added in the
event that the underlying typeface could not create all the mask
formats and because SkRandomScalerContext::generateMetrics used to
calculate the bounds from the path directly. Neither is a current
concern, especially since the test uses a portable typeface that always
draws from paths anyway.

Change-Id: I63662323aa35a5dd000193f2e27f1ca3b70fa637
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/974358
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>