[graphite] Support lifting expressions to vertex shader
This adds the concept of a "liftable expression generator" to
ShaderSnippets, which can be invoked either in the snippet's preamble
(which is normally called in the fragment shader), or directly in the
vertex shader if ShaderInfo logic determines the expression can indeed
be lifted. In the vertex shader the evaluated expression is captured in
a varying which is then used in place of the expression in the preamble.

For now, only top-level local matrix transformations are lifted.

Bug: b/238761058
Change-Id: I7f9761473d860b3c3fddb45f87d96175e1fea133
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/973277
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>